refactor(kilo-chat): split conversation screen logic
