+++ apps/mobile/src/components/kilo-chat/conversation-screen.tsx
@@ -1,70 +1,31 @@
-/* eslint-disable max-lines */
 import { useActionSheet } from '@expo/react-native-action-sheet';
-import * as Clipboard from 'expo-clipboard';
 import * as Haptics from 'expo-haptics';
-import {
-  attemptMarkCurrentConversationRead,
-  clearMarkReadRetry,
-  clearPendingAction,
-  createMarkReadRetryState,
-  createMarkReadState,
-  latestMarkReadMessageId,
-  type PendingAction,
-  tryStartPendingAction,
-  useAddReaction,
-  useBotStatus,
-  useDeleteMessage,
-  useEditMessage,
-  useExecuteAction,
-  useRemoveReaction,
-} from '@kilocode/kilo-chat-hooks';
-import {
-  buildMessageActionAvailability,
-  buildMessageEditContent,
-  contentBlocksToText,
-  type ConversationDetailResponse,
-  type ExecApprovalDecision,
-  formatKiloChatError,
-  type InputContentBlock,
-  type Message,
-} from '@kilocode/kilo-chat';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useBotStatus } from '@kilocode/kilo-chat-hooks';
+import { type ConversationDetailResponse } from '@kilocode/kilo-chat';
+import { useCallback } from 'react';
 import { Alert, View } from 'react-native';
 import { useFocusEffect, useRouter } from 'expo-router';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { toast } from 'sonner-native';
 
-import { QueryError } from '@/components/query-error';
-import { Skeleton } from '@/components/ui/skeleton';
 import { AppAwareKeyboardPaddingView } from './app-aware-keyboard-padding';
 import { ConversationHeader } from './conversation-header';
-import { resolveMobileMessageInputAvailability } from './bot-send-state';
-import { executeActionWithMobileFeedback } from './execute-action-feedback';
-import { buildMessageActionSheetOptions, getSelectedMessageAction } from './message-actions';
+import {
+  ConversationHistoryErrorView,
+  ConversationHistoryLoadingView,
+} from './conversation-history-state-views';
 import { MessageInput } from './message-input';
-import { type MessageInputSubmitControls } from './message-input-state';
 import { MessageList } from './message-list';
 import { MessageReactionPickerSheet } from './message-reaction-picker-sheet';
-import {
-  buildSendMessageVariables,
-  canCopyMessage,
-  canToggleReaction,
-  createSendMessageClientId,
-  getEditableAttachmentBlocks,
-  getVisibleEditableAttachmentBlocks,
-} from './message-presentation';
-import {
-  getMessageHistoryContentState,
-  shouldMarkLatestMessageRead,
-} from './message-history-state';
+import { getMessageHistoryContentState } from './message-history-state';
 import { useConversationPresence } from './hooks/use-conversation-presence';
 import { useConversationEventSubscription } from './hooks/use-conversation-event-subscription';
 import { useLeaveConversation } from './hooks/use-conversations';
 import { useMobileTypingState, useTypingSender } from './hooks/use-typing';
 import { useKiloChatClient } from './hooks/use-kilo-chat-client';
-import { useAppActiveAndFocused } from './hooks/use-app-active-and-focused';
-import { useMarkRead } from './hooks/use-mark-read';
-import { useMessageCacheUpdater, useMessages, useSendMessage } from './hooks/use-messages';
+import { useConversationMarkRead } from './hooks/use-conversation-mark-read';
+import { useConversationMessageController } from './hooks/use-conversation-message-controller';
+import { useMessageCacheUpdater, useMessages } from './hooks/use-messages';
 import { useNowTicker } from './hooks/use-now-ticker';
 import { useCurrentUserId } from './hooks/use-current-user-id';
 import { useAllKiloClawInstances, useInstanceContext } from '@/lib/hooks/use-instance-context';
@@ -85,23 +46,6 @@ type Props = {
   conversationMembers: ConversationDetailResponse['members'];
 };
 
-function editableText(message: Message): string {
-  return message.content
-    .filter(block => block.type === 'text')
-    .map(block => block.text)
-    .join('\n');
-}
-
-function MessageHistorySkeleton() {
-  return (
-    <View className="flex-1 justify-end gap-3 px-4 py-6">
-      <Skeleton className="h-14 w-3/4 rounded-md" />
-      <Skeleton className="ml-auto h-16 w-2/3 rounded-md" />
-      <Skeleton className="h-20 w-5/6 rounded-md" />
-    </View>
-  );
-}
-
 export function ConversationScreen({
   sandboxId,
   conversationId,
@@ -114,14 +58,6 @@ export function ConversationScreen({
   const currentUserId = useCurrentUserId();
   const { showActionSheetWithOptions } = useActionSheet();
   const { bottom } = useSafeAreaInsets();
-  const [editingMessage, setEditingMessage] = useState<Message | null>(null);
-  const [removedEditAttachmentIds, setRemovedEditAttachmentIds] = useState<string[]>([]);
-  const [replyingTo, setReplyingTo] = useState<Message | null>(null);
-  const [reactionPickerMessage, setReactionPickerMessage] = useState<Message | null>(null);
-  const [recentReactions, setRecentReactions] = useState<string[]>([]);
-  const [pendingAction, setPendingAction] = useState<PendingAction | null>(null);
-  const [scrollToNewestRequest, setScrollToNewestRequest] = useState(0);
-  const pendingActionRef = useRef<PendingAction | null>(null);
   const instanceContext = useInstanceContext(sandboxId);
   const instanceStatusQuery = useKiloClawStatus(
     instanceContext.organizationId,
@@ -143,24 +79,13 @@ export function ConversationScreen({
   });
   const hasInitialMessages = messageHistoryState === 'ready';
   const messages = hasInitialMessages ? (messagesQuery.data?.messages ?? []) : [];
-  const latestMessageId = latestMarkReadMessageId(messages);
-  const latestMarkReadMessageSenderId =
-    latestMessageId === null
-      ? null
-      : (messages.find(message => message.id === latestMessageId)?.senderId ?? null);
   const fetchOlder = useCallback(() => {
     if (messagesQuery.hasNextPage && !messagesQuery.isFetchingNextPage) {
       void messagesQuery.fetchNextPage();
     }
   }, [messagesQuery]);
 
-  const sendMutation = useSendMessage(client, conversationId, currentUserId);
   const leaveConversation = useLeaveConversation(client);
-  const editMessage = useEditMessage(client, conversationId);
-  const deleteMessage = useDeleteMessage(client, conversationId);
-  const executeAction = useExecuteAction(client, conversationId, currentUserId);
-  const addReaction = useAddReaction(client, conversationId, currentUserId);
-  const removeReaction = useRemoveReaction(client, conversationId, currentUserId);
   const { typingMembers, clearTypingForMember } = useMobileTypingState({
     client,
     currentUserId,
@@ -168,25 +93,13 @@ export function ConversationScreen({
     conversationId,
   });
   const sendTyping = useTypingSender(client, conversationId);
-  const editingText = useMemo(
-    () => (editingMessage ? editableText(editingMessage) : ''),
-    [editingMessage]
-  );
-  const editingAttachments = useMemo(
-    () => (editingMessage ? getEditableAttachmentBlocks(editingMessage) : []),
-    [editingMessage]
-  );
-  const visibleEditingAttachments = useMemo(
-    () => getVisibleEditableAttachmentBlocks(editingAttachments, removedEditAttachmentIds),
-    [editingAttachments, removedEditAttachmentIds]
-  );
-  const inputAvailability = resolveMobileMessageInputAvailability({
+  const messageController = useConversationMessageController({
+    client,
+    conversationId,
     currentUserId,
     instanceStatus,
     presence: botPresence,
     now,
-    pendingMutation: sendMutation.isPending || editMessage.isPending,
-    editing: editingMessage !== null,
   });
 
   const canSwitchInstance = (instances?.length ?? 0) > 1;
@@ -244,234 +157,6 @@ export function ConversationScreen({
     sandboxId,
     showActionSheetWithOptions,
   ]);
-  const handleSend = useCallback(
-    (
-      content: InputContentBlock[],
-      inReplyToMessageId?: string,
-      controls?: MessageInputSubmitControls
-    ) => {
-      if (!editingMessage && inputAvailability.disabled) {
-        return;
-      }
-      if (editingMessage) {
-        const editContent = buildMessageEditContent({
-          text: contentBlocksToText(content),
-          originalAttachments: editingAttachments,
-          removedAttachmentIds: removedEditAttachmentIds,
-        });
-        editMessage.mutate(
-          {
-            messageId: editingMessage.id,
-            conversationId,
-            content: editContent,
-            timestamp: Date.now(),
-          },
-          {
-            onSuccess: () => {
-              controls?.clearDraft();
-              setEditingMessage(null);
-              setRemovedEditAttachmentIds([]);
-              void Haptics.notificationAsync(Haptics.NotificationFeedbackType.Success);
-            },
-            onError: err => {
-              toast.error(formatKiloChatError(err, 'Failed to edit message'));
-            },
-          }
-        );
-        return;
-      }
-      sendMutation.mutate(
-        buildSendMessageVariables({
-          conversationId,
-          content,
-          clientId: createSendMessageClientId(),
-          inReplyToMessageId,
-        }),
-        {
-          onSuccess: () => {
-            if (controls?.clearDraft() ?? false) {
-              setReplyingTo(null);
-            }
-            void Haptics.notificationAsync(Haptics.NotificationFeedbackType.Success);
-          },
-          onError: err => {
-            toast.error(formatKiloChatError(err, 'Failed to send message'));
-          },
-        }
-      );
-      setScrollToNewestRequest(request => request + 1);
-    },
-    [
-      conversationId,
-      editMessage,
-      editingAttachments,
-      editingMessage,
-      inputAvailability.disabled,
-      removedEditAttachmentIds,
-      sendMutation,
-    ]
-  );
-  const handleRemoveEditableAttachment = useCallback((attachmentId: string) => {
-    setRemovedEditAttachmentIds(current =>
-      current.includes(attachmentId) ? current : [...current, attachmentId]
-    );
-  }, []);
-  const handleReactionPress = useCallback(
-    (message: Message, emoji: string) => {
-      if (!currentUserId || !canToggleReaction(message, currentUserId)) {
-        return;
-      }
-      const hasReacted =
-        message.reactions.find(r => r.emoji === emoji)?.memberIds.includes(currentUserId) ?? false;
-      if (hasReacted) {
-        removeReaction.mutate(
-          { messageId: message.id, emoji },
-          {
-            onError: err => {
-              toast.error(formatKiloChatError(err, 'Failed to remove reaction'));
-            },
-          }
-        );
-      } else {
-        addReaction.mutate(
-          { messageId: message.id, emoji },
-          {
-            onError: err => {
-              toast.error(formatKiloChatError(err, 'Failed to add reaction'));
-            },
-          }
-        );
-      }
-      setRecentReactions(previous => [emoji, ...previous.filter(reaction => reaction !== emoji)]);
-      void Haptics.selectionAsync();
-    },
-    [addReaction, currentUserId, removeReaction]
-  );
-  const handleCopyMessage = useCallback(async (message: Message) => {
-    try {
-      await Clipboard.setStringAsync(contentBlocksToText(message.content));
-      toast.success('Copied');
-    } catch {
-      toast.error('Failed to copy');
-    }
-  }, []);
-  const handleExecuteAction = useCallback(
-    (message: Message, groupId: string, value: ExecApprovalDecision) => {
-      const nextPendingAction = { messageId: message.id, groupId };
-      if (!tryStartPendingAction(pendingActionRef, nextPendingAction)) {
-        return;
-      }
-      setPendingAction(pendingActionRef.current);
-      executeActionWithMobileFeedback({
-        executeAction,
-        message,
-        groupId,
-        value,
-        onSettled: () => {
-          clearPendingAction(pendingActionRef, nextPendingAction);
-          setPendingAction(pendingActionRef.current);
-        },
-      });
-    },
-    [executeAction]
-  );
-  const handleLongPressMessage = useCallback(
-    (message: Message) => {
-      const isOwnMessage = currentUserId !== null && message.senderId === currentUserId;
-      const actionAvailability = buildMessageActionAvailability(message, isOwnMessage);
-      const isPendingMessage = message.id.startsWith('pending-');
-      const actionSheet = buildMessageActionSheetOptions({
-        canReact: currentUserId !== null && actionAvailability.canReact,
-        canReply: actionAvailability.canReply,
-        canCopy: canCopyMessage(message),
-        canEdit: actionAvailability.canEdit,
-        canDelete: actionAvailability.canDelete,
-        isPendingMessage,
-      });
-      showActionSheetWithOptions(
-        {
-          options: actionSheet.options,
-          cancelButtonIndex: actionSheet.cancelButtonIndex,
-          destructiveButtonIndex: actionSheet.destructiveButtonIndex,
-          title: 'Message actions',
-          containerStyle: { paddingBottom: bottom },
-        },
-        index => {
-          const selectedAction = getSelectedMessageAction(actionSheet, index);
-          if (!selectedAction) {
-            return;
-          }
-
-          if (selectedAction.kind === 'reaction') {
-            handleReactionPress(message, selectedAction.emoji);
-            return;
-          }
-          if (selectedAction.kind === 'more-reactions') {
-            setReactionPickerMessage(message);
-            return;
-          }
-          if (selectedAction.kind === 'copy') {
-            void handleCopyMessage(message);
-            return;
-          }
-          if (selectedAction.kind === 'reply') {
-            setEditingMessage(null);
-            setRemovedEditAttachmentIds([]);
-            setReplyingTo(message);
-            return;
-          }
-          if (selectedAction.kind === 'edit') {
-            setReplyingTo(null);
-            setEditingMessage(message);
-            setRemovedEditAttachmentIds([]);
-            return;
-          }
-
-          Alert.alert('Delete message?', 'This will remove the message from the conversation.', [
-            { text: 'Cancel', style: 'cancel' },
-            {
-              text: 'Delete',
-              style: 'destructive',
-              onPress: () => {
-                deleteMessage.mutate(
-                  { messageId: message.id, conversationId },
-                  {
-                    onError: err => {
-                      toast.error(formatKiloChatError(err, 'Failed to delete message'));
-                    },
-                  }
-                );
-              },
-            },
-          ]);
-        }
-      );
-    },
-    [
-      bottom,
-      conversationId,
-      currentUserId,
-      deleteMessage,
-      handleCopyMessage,
-      handleReactionPress,
-      showActionSheetWithOptions,
-    ]
-  );
-  const handleSwipeReplyMessage = useCallback(
-    (message: Message) => {
-      const isOwnMessage = currentUserId !== null && message.senderId === currentUserId;
-      const actionAvailability = buildMessageActionAvailability(message, isOwnMessage);
-      if (!actionAvailability.canReply) {
-        return;
-      }
-      setEditingMessage(null);
-      setRemovedEditAttachmentIds([]);
-      setReplyingTo(message);
-      void Haptics.selectionAsync();
-    },
-    [currentUserId]
-  );
-
   useConversationPresence(sandboxId, conversationId);
   useConversationEventSubscription(sandboxId, conversationId);
   const handleActionFailed = useCallback(() => {
@@ -488,83 +173,14 @@ export function ConversationScreen({
     handleActionFailed,
     handleMessageDeliveryFailed
   );
-
-  const activeAndFocused = useAppActiveAndFocused();
-  const markRead = useMarkRead(client);
-  const markReadStateRef = useRef(createMarkReadState());
-  const markReadRetryStateRef = useRef(createMarkReadRetryState());
-  const currentMarkReadMarker =
-    latestMessageId === null ? null : `${conversationId}:${latestMessageId}`;
-  const currentMarkReadMarkerRef = useRef<string | null>(currentMarkReadMarker);
-  const activeAndFocusedRef = useRef(activeAndFocused);
-  const markCurrentConversationReadRef = useRef<(() => void) | null>(null);
-  currentMarkReadMarkerRef.current = currentMarkReadMarker;
-  activeAndFocusedRef.current = activeAndFocused;
-
-  const markCurrentConversationRead = useCallback(() => {
-    if (!hasInitialMessages || latestMessageId === null || currentMarkReadMarker === null) {
-      return;
-    }
-    if (
-      !shouldMarkLatestMessageRead({
-        currentUserId,
-        latestMessageSenderId: latestMarkReadMessageSenderId,
-      })
-    ) {
-      return;
-    }
-    const marker = currentMarkReadMarker;
-    void attemptMarkCurrentConversationRead({
-      marker,
-      markReadState: markReadStateRef.current,
-      retryState: markReadRetryStateRef.current,
-      currentMarker: () => currentMarkReadMarkerRef.current,
-      isActive: () => activeAndFocusedRef.current,
-      markRead: async () => {
-        await markRead(sandboxId, conversationId, latestMessageId);
-      },
-      retry: () => {
-        markCurrentConversationReadRef.current?.();
-      },
-    });
-  }, [
+  useConversationMarkRead({
+    client,
     conversationId,
-    currentMarkReadMarker,
     currentUserId,
     hasInitialMessages,
-    latestMessageId,
-    latestMarkReadMessageSenderId,
-    markRead,
+    messages,
     sandboxId,
-  ]);
-  markCurrentConversationReadRef.current = markCurrentConversationRead;
-
-  useEffect(() => {
-    if (!activeAndFocused || currentMarkReadMarker === null) {
-      clearMarkReadRetry(markReadRetryStateRef.current);
-      return;
-    }
-    if (
-      markReadRetryStateRef.current.marker !== null &&
-      markReadRetryStateRef.current.marker !== currentMarkReadMarker
-    ) {
-      clearMarkReadRetry(markReadRetryStateRef.current);
-    }
-  }, [activeAndFocused, currentMarkReadMarker]);
-
-  useEffect(() => {
-    const retryState = markReadRetryStateRef.current;
-    return () => {
-      clearMarkReadRetry(retryState);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!activeAndFocused) {
-      return;
-    }
-    markCurrentConversationRead();
-  }, [activeAndFocused, markCurrentConversationRead]);
+  });
 
   useFocusEffect(
     useCallback(() => {
@@ -576,30 +192,18 @@ export function ConversationScreen({
   );
 
   if (messageHistoryState === 'loading') {
-    return (
-      <View className="flex-1">
-        <ConversationHeader title={conversationTitle} subtitle={instanceLabel} />
-        <AppAwareKeyboardPaddingView className="flex-1">
-          <MessageHistorySkeleton />
-        </AppAwareKeyboardPaddingView>
-      </View>
-    );
+    return <ConversationHistoryLoadingView title={conversationTitle} subtitle={instanceLabel} />;
   }
 
   if (messageHistoryState === 'error') {
     return (
-      <View className="flex-1">
-        <ConversationHeader title={conversationTitle} subtitle={instanceLabel} />
-        <AppAwareKeyboardPaddingView className="flex-1">
-          <QueryError
-            className="flex-1"
-            message="Could not load conversation history"
-            onRetry={() => {
-              void messagesQuery.refetch();
-            }}
-          />
-        </AppAwareKeyboardPaddingView>
-      </View>
+      <ConversationHistoryErrorView
+        title={conversationTitle}
+        subtitle={instanceLabel}
+        onRetry={() => {
+          void messagesQuery.refetch();
+        }}
+      />
     );
   }
 
@@ -622,59 +226,59 @@ export function ConversationScreen({
           botName={instanceLabel}
           fetchOlder={fetchOlder}
           isFetchingOlder={messagesQuery.isFetchingNextPage}
-          pendingAction={pendingAction}
-          scrollToNewestRequest={scrollToNewestRequest}
-          onExecuteAction={handleExecuteAction}
-          onLongPressMessage={handleLongPressMessage}
-          onSwipeReplyMessage={handleSwipeReplyMessage}
-          onReactionPress={handleReactionPress}
+          pendingAction={messageController.pendingAction}
+          scrollToNewestRequest={messageController.scrollToNewestRequest}
+          onExecuteAction={messageController.handleExecuteAction}
+          onLongPressMessage={messageController.handleLongPressMessage}
+          onSwipeReplyMessage={messageController.handleSwipeReplyMessage}
+          onReactionPress={messageController.handleReactionPress}
         />
         <MessageInput
-          key={editingMessage?.id ?? 'compose'}
-          onSend={handleSend}
+          key={messageController.editingMessage?.id ?? 'compose'}
+          onSend={messageController.handleSend}
           onTyping={sendTyping}
           client={client}
           conversationId={conversationId}
           hasAttachmentsCapability={hasAttachmentsCapability}
-          disabled={inputAvailability.disabled}
-          submitDisabled={inputAvailability.submitDisabled}
-          disabledReason={inputAvailability.disabledReason}
-          initialText={editingText}
-          isEditing={editingMessage !== null}
-          editableAttachments={visibleEditingAttachments}
-          onRemoveEditableAttachment={handleRemoveEditableAttachment}
+          disabled={messageController.inputAvailability.disabled}
+          submitDisabled={messageController.inputAvailability.submitDisabled}
+          disabledReason={messageController.inputAvailability.disabledReason}
+          initialText={messageController.editingText}
+          isEditing={messageController.editingMessage !== null}
+          editableAttachments={messageController.visibleEditingAttachments}
+          onRemoveEditableAttachment={messageController.handleRemoveEditableAttachment}
           botName={instanceLabel}
           typingMembers={typingMembers}
-          replyingTo={replyingTo}
+          replyingTo={messageController.replyingTo}
           onCancelReply={
-            replyingTo
+            messageController.replyingTo
               ? () => {
-                  setReplyingTo(null);
+                  messageController.setReplyingTo(null);
                 }
               : undefined
           }
           onCancelEdit={
-            editingMessage
+            messageController.editingMessage
               ? () => {
-                  setEditingMessage(null);
-                  setRemovedEditAttachmentIds([]);
+                  messageController.setEditingMessage(null);
+                  messageController.setRemovedEditAttachmentIds([]);
                 }
               : undefined
           }
         />
       </AppAwareKeyboardPaddingView>
       <MessageReactionPickerSheet
-        visible={reactionPickerMessage !== null}
-        recentReactions={recentReactions}
+        visible={messageController.reactionPickerMessage !== null}
+        recentReactions={messageController.recentReactions}
         onClose={() => {
-          setReactionPickerMessage(null);
+          messageController.setReactionPickerMessage(null);
         }}
         onSelect={emoji => {
-          const message = reactionPickerMessage;
+          const message = messageController.reactionPickerMessage;
           if (message) {
-            handleReactionPress(message, emoji);
+            messageController.handleReactionPress(message, emoji);
           }
-          setReactionPickerMessage(null);
+          messageController.setReactionPickerMessage(null);
         }}
       />
     </View>
